requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.1
Django==4.2.5
django-ckeditor==6.7.0
django-filter==23.3
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
mccabe==0.7.0
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
